refactor section-title + /components exports
