+++ src/components/home-page/index.ts
@@ -1,4 +1,5 @@
 export * from './about';
 export * from './coming-soon';
+export * from "./contact"
 export * from './hero';
 export * from './services';
